+++ public/app/plugins/datasource/grafana-azure-monitor-datasource/components/ResourcePicker/EntryIcon.tsx
@@ -1,0 +1,31 @@
+import { Icon } from '@grafana/ui';
+import React, { useCallback, useEffect, useState } from 'react';
+
+import { ResourceRow, ResourceRowType } from './types';
+
+interface EntryIconProps {
+  entry: ResourceRow;
+  isOpen: boolean;
+}
+
+export const EntryIcon: React.FC<EntryIconProps> = ({ isOpen, entry: { type } }) => {
+  switch (type) {
+    case ResourceRowType.Subscription:
+      return <Icon name="layer-group" />;
+
+    case ResourceRowType.ResourceGroup:
+      return <Icon name={isOpen ? 'folder-open' : 'folder'} />;
+
+    case ResourceRowType.Resource:
+      return <Icon name="cube" />;
+
+    case ResourceRowType.VariableGroup:
+      return <Icon name="x" />;
+
+    case ResourceRowType.Variable:
+      return <Icon name="x" />;
+
+    default:
+      return null;
+  }
+};
