+++ public/app/plugins/datasource/grafana-azure-monitor-datasource/components/ResourcePicker/NestedEntry.test.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+describe('NestedEntry', () => {
+  it('should be selectable', () => {});
+});
